8 nes

+++ 0x02-functions_nested_loops/main.h
@@ -1,4 +1,4 @@
-int print_last_digit(int);
+int print_last_digit(long);
 int _abs(int);
 int _putchar(char c);
 void print_alphabet();
